+++ docs/design/tools/transfer_channel_benchmark/README.md
@@ -1,0 +1,1 @@
+../../../../lmcache/tools/transfer_channel_benchmark/README.md
