+++ lmcache/tools/transfer_channel_benchmark/__main__.py
@@ -6,10 +6,10 @@ import argparse
 import sys
 
 # First Party
-from lmcache.tools.transfer_channel_benchmark.benchmark import (
-    add_benchmark_arguments,
-    run_benchmark,
-)
+# Only the torch-free config module is imported eagerly so that `-h` works
+# without torch installed. The runtime (which needs torch) is imported after
+# arguments are parsed.
+from lmcache.tools.transfer_channel_benchmark.config import add_benchmark_arguments
 
 
 def main() -> int:
@@ -25,6 +25,13 @@ def main() -> int:
     )
     add_benchmark_arguments(parser)
     args = parser.parse_args()
+
+    # Imported here (not at module load) so this entry point can show --help
+    # without requiring torch; importing it raises a clear error if torch is
+    # missing.
+    # First Party
+    from lmcache.tools.transfer_channel_benchmark.benchmark import run_benchmark
+
     return 0 if run_benchmark(args) else 1
